Finish fixing tests, code org, dockerfile updates
