+++ pyoutline/tests/__init__.py
@@ -12,5 +12,4 @@
 #  See the License for the specific language governing permissions and
 #  limitations under the License.
 
-import depend
-import session
+import test_utils
